Fix to #13511 - SelectMany with DefaultIfEmpty() generates unexpected INNER JOIN
Problem was that for navigations inside SelectMany we would always produce inner joins, however, when the navigation chain ends in DefaultIfEmpty, we can suspect that SelectMany operation is supposed to return at least one result per source element. For that purpose we need to produce LOJ.

Fix is to check if the navigation chain inside SelectMany ends in DefaultIfEmpty, and if so - generate LOJ calls instead.